Set RC ip to just after the hook address
Otherwise, setting a breakpoint in a function that is short will fail
and make the unwinder think we're in the function after instead.
This is still not great, and our Hooking asm template is too long.
It can probably be made slightly better using extender pages, but
really we need a one-byte solution.

Fixes https://github.com/Keno/Gallium.jl/issues/88