Fix #89
When a breakpoint on error occurs, we're not in fact at the stacktop in the
first frame (since we skip the backtrace frame). Further, since this is
a noreturn callsite, the ip might be after the end of the function if
it's not adjusted (which it isn't if we're at the stacktop).