Fix recursive breakpoint_on_error
Really inference should stop using errors as part of its normal operation,
but until it does, we need to play whack-a-mole to make sure inference doesn't
get triggered until we unhook the breakpoint. We do this here by moving
the unhooking earlier. These kinds of errors are very tricky to debug.
Luckily rr has come far enough to handle this.

Also disable the tests for xmm recover, because we seem to no longer have
that debug info for some reason.

Fixes #151, #120